+++ package/multimedia/aml_libs/src/audio_codec/libcook/rm_parser_internal.h
@@ -109,7 +109,7 @@ struct rm_properties_hdr
     UINT32 index_offset; /* offset to beginning of index data */
     UINT32 data_offset;  /* offset to beginning of packet data */
     UINT16 num_streams;  /* total number of streams in clip */
-    UINT16 flags;        /* various and sundry */
+    UINT32 flags;        /* various and sundry */
 };
 
 /* rm content header -- title, author, copyright */
